Organize editor folder structure 🚚
