+++ src/edit/Toolbar/CaptionUpload.vue
@@ -26,7 +26,7 @@
 
 <script>
 import ConfirmationDialog from './ConfirmationDialog.vue';
-import { ELEMENT_STATE } from '../shared';
+import { ELEMENT_STATE } from '../../shared';
 import UploadBtn from './UploadBtn.vue';
 
 export default {
