+++ src/edit/Toolbar/PosterFrameUpload/PosterFrame.vue
